defrag plan/*.go
