+++ plan/plan.go
@@ -1,206 +1,22 @@
 package plan
 
 import (
-	"bytes"
 	"context"
+
 	"fmt"
-	"slices"
-	"strings"
 
 	blueprintPkg "github.com/fornellas/resonance/blueprint"
-	"github.com/fornellas/resonance/diff"
 	"github.com/fornellas/resonance/host/types"
 	"github.com/fornellas/resonance/log"
 	resourcesPkg "github.com/fornellas/resonance/resources"
+	storePkg "github.com/fornellas/resonance/store"
 )
-
-// ResourceDiff holds the diff for applying each Action resource.
-type ResourceDiff struct {
-	// An emoji representing the action
-	Emoji rune
-	// The resource Id
-	Id string
-	// The diff to apply the resource.
-	Chunks diff.Chunks
-}
-
-func (r *ResourceDiff) String() string {
-	return fmt.Sprintf("%c%s", r.Emoji, r.Id)
-}
-
-// NewResourceDiff creates a new ResourceDiff
-func NewResourceDiff(emoji rune, planResource resourcesPkg.Resource, chunks diff.Chunks) *ResourceDiff {
-	return &ResourceDiff{
-		Emoji:  emoji,
-		Id:     resourcesPkg.GetResourceId(planResource),
-		Chunks: chunks,
-	}
-}
-
-// Action holds actions required to apply, calculated from a Step.
-type Action struct {
-	// A string with the resource type (eg: File, APTPackage)
-	ResourceType string
-	// The calculated diff for each Step resource when applying it.
-	ResourceDiffs []*ResourceDiff
-	// The resources that needs to be applied for the Step (eg: if no changes required for the
-	// resource, it won't be here)
-	ApplyResources resourcesPkg.Resources
-}
-
-// NewAction creates a new Action for a given Step, calculating required changes as a function
-// of the before (apply) state of the step resources.
-func NewAction(step *blueprintPkg.Step, beforeResourceMap resourcesPkg.ResourceMap) *Action {
-	action := &Action{
-		ResourceType:   step.Type(),
-		ResourceDiffs:  make([]*ResourceDiff, len(step.Resources())),
-		ApplyResources: resourcesPkg.Resources{},
-	}
-	for j, planResource := range step.Resources() {
-		beforeResource := beforeResourceMap.GetResourceWithSameTypeId(planResource)
-		if beforeResource == nil {
-			panic("bug: before resource not found")
-		}
-		var resourceAction *ResourceDiff
-		if resourcesPkg.Satisfies(beforeResource, planResource) {
-			resourceAction = NewResourceDiff('✅', planResource, nil)
-		} else {
-			action.ApplyResources = append(action.ApplyResources, planResource)
-			var emoji rune
-			if resourcesPkg.GetResourceAbsent(beforeResource) {
-				emoji = '🔧'
-			} else {
-				if resourcesPkg.GetResourceAbsent(planResource) {
-					emoji = '🗑'
-				} else {
-					emoji = '🔄'
-				}
-			}
-			resourceAction = NewResourceDiff(emoji, planResource, diff.DiffAsYaml(
-				beforeResource, planResource,
-			))
-		}
-		action.ResourceDiffs[j] = resourceAction
-	}
-	slices.SortFunc(action.ResourceDiffs, func(a, b *ResourceDiff) int {
-		return strings.Compare(a.Id, b.Id)
-	})
-	slices.SortFunc(action.ApplyResources, func(a, b resourcesPkg.Resource) int {
-		return strings.Compare(resourcesPkg.GetResourceId(a), resourcesPkg.GetResourceId(b))
-	})
-	return action
-}
-
-// String returns a single-line representation of the action.
-func (a *Action) String() string {
-	actionStrs := make([]string, len(a.ResourceDiffs))
-	for i, resourceDiff := range a.ResourceDiffs {
-		actionStrs[i] = resourceDiff.String()
-	}
-	return fmt.Sprintf("%s:%s", a.ResourceType, strings.Join(actionStrs, ","))
-}
-
-// DiffString returns details on the diff required to apply this action.
-func (a *Action) DiffString() string {
-	var buff bytes.Buffer
-
-	if len(a.ResourceDiffs) == 1 {
-		resourceDiffs := a.ResourceDiffs[0]
-		if len(resourceDiffs.Chunks) > 0 {
-			fmt.Fprintf(&buff, "%s", resourceDiffs.Chunks.String())
-		}
-	} else {
-		for _, resourceDiffs := range a.ResourceDiffs {
-			if len(resourceDiffs.Chunks) > 0 {
-				fmt.Fprintf(&buff, "%s:\n", resourceDiffs.Id)
-				fmt.Fprintf(&buff, "  %s\n",
-					strings.Join(
-						strings.Split(
-							strings.TrimSuffix(resourceDiffs.Chunks.String(), "\n"),
-							"\n",
-						),
-						"\n  ",
-					),
-				)
-			}
-		}
-	}
-
-	return strings.TrimSuffix(buff.String(), "\n")
-}
-
-// DetailedString returns a multi-line string, fully describing the action and its diff.
-func (a *Action) DetailedString() string {
-	diffStr := a.DiffString()
-	if len(diffStr) > 0 {
-		return strings.TrimSuffix(
-			fmt.Sprintf("%s\n  %s\n",
-				a,
-				strings.Join(
-					strings.Split(
-						strings.TrimSuffix(diffStr, "\n"),
-						"\n",
-					),
-					"\n  ",
-				),
-			),
-			"\n",
-		)
-	} else {
-		return a.String()
-	}
-}
-
-// Apply commits all required changes for action to given Host.
-func (a *Action) Apply(ctx context.Context, host types.Host) error {
-	ctx, logger := log.WithGroup(ctx, a.String())
-	diffStr := a.DiffString()
-	if len(diffStr) > 0 {
-		ctx, logger = log.WithAttrs(ctx, "diff", diffStr)
-		logger.Info("Applying changes")
-	} else {
-		logger.Info("Nothing to do")
-	}
-
-	if len(a.ApplyResources) == 0 {
-		return nil
-	}
-
-	isGroupResource := resourcesPkg.IsGroupResource(
-		resourcesPkg.GetResourceTypeName(a.ApplyResources[0]),
-	)
-
-	if isGroupResource {
-		groupResource := resourcesPkg.GetGroupResourceByTypeName(a.ResourceType)
-		if groupResource == nil {
-			panic("bug: bad GroupResource")
-		}
-		if err := groupResource.Apply(ctx, host, a.ApplyResources); err != nil {
-			return err
-		}
-	} else {
-		if len(a.ApplyResources) != 1 {
-			panic("bug: can't have more than one SingleResource")
-		}
-		singleResource, ok := a.ApplyResources[0].(resourcesPkg.SingleResource)
-		if !ok {
-			panic("bug: is not SingleResource")
-		}
-		if err := singleResource.Apply(ctx, host); err != nil {
-			return err
-		}
-	}
-
-	return nil
-}
 
 // Plan holds all actions required to apply changes to a host.
 // This enables evaluating all changes before they are applied.
 type Plan []*Action
 
-// NewPlan crafts a new plan as a function of: the delined target state, the last state (how the
-// host is at the moment) and the original state of any managed resource.
-func NewPlan(
+func compilePlan(
 	ctx context.Context,
 	targetBlueprint, lastBlueprint *blueprintPkg.Blueprint,
 	loadOriginalResource func(ctx context.Context, resource resourcesPkg.Resource) (resourcesPkg.Resource, error),
@@ -280,6 +96,204 @@ func NewPlan(
 		plan[i] = NewAction(step, beforeResourceMap)
 	}
 	return plan, nil
+}
+
+// createTargetBlueprint crafts a new target Blueprint from given unresolved target
+// resources, against given host.
+// The blueprint is saved at given store.
+func createTargetBlueprint(
+	ctx context.Context,
+	host types.Host,
+	targetResources resourcesPkg.Resources,
+) (*blueprintPkg.Blueprint, error) {
+	ctx, _ = log.WithGroup(ctx, "🎯 Crafting target Blueprint")
+
+	var targetBlueprint *blueprintPkg.Blueprint
+	{
+		var err error
+		targetBlueprint, err = blueprintPkg.NewBlueprintFromResources(ctx, targetResources)
+		if err != nil {
+			return nil, err
+		}
+		{
+			ctx, _ := log.WithGroup(ctx, "⚙️ Resolving target Blueprint")
+			if err := targetBlueprint.Resolve(ctx, host); err != nil {
+				return nil, err
+			}
+		}
+		{
+			_, logger := log.WithGroup(ctx, "🧩 Crafted target Blueprint")
+			for _, step := range targetBlueprint.Steps {
+				resources := step.Resources()
+				if len(resources) == 1 {
+					logger.Info(resourcesPkg.GetResourceTypeName(resources[0]), "yaml", resourcesPkg.GetResourceYaml(resources[0]))
+				} else {
+					logger.Info(resourcesPkg.GetResourceTypeName(resources[0]), "yaml", resourcesPkg.GetResourcesYaml(resources))
+				}
+			}
+		}
+	}
+
+	return targetBlueprint, nil
+}
+
+// saveOriginalResourcesState, for each resource from target Blueprint, loads the
+// resource current state from given host, and saves it at store as the original
+// state.
+func saveOriginalResourcesState(
+	ctx context.Context,
+	host types.Host,
+	store storePkg.Store,
+	targetBlueprint *blueprintPkg.Blueprint,
+) error {
+	ctx, logger := log.WithGroup(ctx, "🌱 Storing original resource states")
+	for _, step := range targetBlueprint.Steps {
+		noOriginalResources := resourcesPkg.Resources{}
+		for _, resource := range step.Resources() {
+			resource = resourcesPkg.NewResourceWithSameId(resource)
+			hasOriginal, err := store.HasOriginalResource(ctx, resource)
+			if err != nil {
+				return err
+			}
+			if !hasOriginal {
+				noOriginalResources = append(noOriginalResources, resource)
+			}
+		}
+		if len(noOriginalResources) == 0 {
+			continue
+		}
+		var noOriginalStep *blueprintPkg.Step
+		if step.IsSingleResource() {
+			if len(noOriginalResources) != 1 {
+				panic("bug: multiple single resource")
+			}
+			noOriginalStep = blueprintPkg.NewSingleResourceStep(noOriginalResources[0].(resourcesPkg.SingleResource))
+		} else if step.IsGroupResource() {
+			noOriginalStep = blueprintPkg.NewGroupResourceStep(step.MustGroupResource())
+			for _, noOriginalResource := range noOriginalResources {
+				noOriginalStep.AppendGroupResource(noOriginalResource)
+			}
+		} else {
+			panic("bug: invalid step type")
+		}
+		originalStep, err := noOriginalStep.Load(ctx, host)
+		if err != nil {
+			return err
+		}
+
+		originalResources := originalStep.Resources()
+		if len(originalResources) == 1 {
+			logger.Info(resourcesPkg.GetResourceTypeName(originalResources[0]), "yaml", resourcesPkg.GetResourceYaml(originalResources[0]))
+		} else {
+			logger.Info(resourcesPkg.GetResourceTypeName(originalResources[0]), "yaml", resourcesPkg.GetResourcesYaml(originalResources))
+		}
+
+		for _, originalResource := range originalResources {
+			if err := store.SaveOriginalResource(ctx, originalResource); err != nil {
+				return err
+			}
+		}
+	}
+	return nil
+}
+
+// loadOrCreateAndSaveLastBlueprintWithValidation attempts to load last blueprint from store.
+// If available, it validates whether the current host state matches that.
+// If not available, it loads current host state for all resources from given target Blueprint
+// and saves it.
+func loadOrCreateAndSaveLastBlueprintWithValidation(
+	ctx context.Context,
+	host types.Host,
+	store storePkg.Store,
+	targetBlueprint *blueprintPkg.Blueprint,
+) (*blueprintPkg.Blueprint, error) {
+	ctx, logger := log.WithGroup(ctx, "↩️ Loading last Blueprint")
+	lastBlueprint, err := store.LoadLastBlueprint(ctx)
+	if err != nil {
+		return nil, err
+	}
+	if lastBlueprint == nil {
+		logger.Info("🔎 No last Blueprint, loading current state")
+		var err error
+		lastBlueprint, err = targetBlueprint.Load(ctx, host)
+		if err != nil {
+			return nil, err
+		}
+
+		{
+			_, logger := log.WithGroup(ctx, "🧩 Loaded Blueprint")
+			for _, step := range lastBlueprint.Steps {
+				resources := step.Resources()
+				if len(resources) == 1 {
+					logger.Info(resourcesPkg.GetResourceTypeName(resources[0]), "yaml", resourcesPkg.GetResourceYaml(resources[0]))
+				} else {
+					logger.Info(resourcesPkg.GetResourceTypeName(resources[0]), "yaml", resourcesPkg.GetResourcesYaml(resources))
+				}
+			}
+		}
+
+		logger.Info("💾 Saving as last Blueprint")
+		if err := store.SaveLastBlueprint(ctx, lastBlueprint); err != nil {
+			return nil, err
+		}
+	} else {
+		logger.Info("🔎 Validating previous host state")
+		currentBlueprint, err := lastBlueprint.Load(ctx, host)
+		if err != nil {
+			return nil, err
+		}
+
+		chunks := currentBlueprint.Satisfies(lastBlueprint)
+		if chunks.HaveChanges() {
+			return nil, fmt.Errorf(
+				"host state has changed:\n%s",
+				chunks.String(),
+			)
+		}
+	}
+	return lastBlueprint, nil
+}
+
+// CraftPlan crafts a new plan as a function of given targetResources (desired state). It uses
+// given store to persist the original state of all target resources (enabling them to be restored
+// when not managed anymore). If a previous plan was previously applied, its state (last Blueprint)
+// is loaded form store and validated (ensuring we're starting from a clean state). If no last
+// Blueprint stored, then it loads the current state of all targetResources and save it as last
+// Blueprint (enabling rollbacks).
+func CraftPlan(
+	ctx context.Context,
+	host types.Host,
+	store storePkg.Store,
+	targetResources resourcesPkg.Resources,
+) (Plan, *blueprintPkg.Blueprint, *blueprintPkg.Blueprint, error) {
+	targetBlueprint, err := createTargetBlueprint(ctx, host, targetResources)
+	if err != nil {
+		return nil, nil, nil, err
+	}
+
+	err = saveOriginalResourcesState(ctx, host, store, targetBlueprint)
+	if err != nil {
+		return nil, nil, nil, err
+	}
+
+	lastBlueprint, err := loadOrCreateAndSaveLastBlueprintWithValidation(
+		ctx, host, store, targetBlueprint,
+	)
+	if err != nil {
+		return nil, nil, nil, err
+	}
+
+	plan, err := compilePlan(
+		ctx,
+		targetBlueprint,
+		lastBlueprint,
+		store.LoadOriginalResource,
+	)
+	if err != nil {
+		return nil, nil, nil, err
+	}
+
+	return plan, targetBlueprint, lastBlueprint, nil
 }
 
 // Apply commits all required changes for plan to given Host.
